src/requirements.txt
Flask==1.1.2
pymongo==3.10.1
python-dotenv==0.20.0
